added a timer to the loading

+++ 3D_Engine/Timer.cpp
@@ -27,7 +27,7 @@ void Timer::Stop()
 }
 
 // ---------------------------------------------
-float Timer::Read()
+Uint32 Timer::Read()
 {
 	if(running == true)
 	{
